react: refactor front end components based on platform section: activity, portal, or page
