+++ react/src/components/elements/activity/ActivityInstruction.module.scss
