+++ react/src/components/elements/common/DropdownSelector.module.scss
