+++ react/src/components/elements/common/NavBar.module.scss
@@ -56,7 +56,7 @@
 }
 
 .menuLogo {
-  background-image: url("../../tagbull.svg");
+  background-image: url("../../../tagbull.svg");
   background-repeat: no-repeat;
   background-size: contain;
   width: 170px;
