+++ react/src/components/elements/common/ProgressBar.module.scss
