+++ react/src/components/elements/common/ProgressBar.tsx
