+++ react/src/components/elements/index.js
@@ -1,18 +1,18 @@
-export { default as ActivityAction } from "./ActivityAction";
-export { default as ActivityActionButton } from "./ActivityActionButton";
-export { default as ActivityInstruction } from "./ActivityInstruction";
-export { default as BigButton } from "./BigButton";
-export { default as ButtonSelector } from "./ButtonSelector";
-export { default as Completion } from "./Completion";
-export { default as DropdownSelector } from "./DropdownSelector";
-export { default as Footer } from "./Footer";
-export { default as HelpButton } from "./HelpButton";
-export { default as ImageUpload } from "./ImageUpload";
-export { default as LoginForm } from "./LoginForm";
-export { default as NavBar } from "./NavBar";
-export { default as ProgressBar } from "./ProgressBar";
-export { default as ProjectsList } from "./ProjectsList";
-export { default as SamplePreview } from "./SamplePreview";
-export { default as TagDiffPreview } from "./TagDiffPreview";
-export { default as TagPreview } from "./TagPreview";
-export { default as Welcome } from "./Welcome";
+export { default as ActivityAction } from "./activity/ActivityAction";
+export { default as ActivityActionButton } from "./activity/ActivityActionButton";
+export { default as ActivityInstruction } from "./activity/ActivityInstruction";
+export { default as BigButton } from "./common/BigButton";
+export { default as ButtonSelector } from "./common/ButtonSelector";
+export { default as Completion } from "./portal/Completion";
+export { default as DropdownSelector } from "./common/DropdownSelector";
+export { default as Footer } from "./common/Footer";
+export { default as HelpButton } from "./activity/HelpButton";
+export { default as ImageUpload } from "./portal/ImageUpload";
+export { default as LoginForm } from "./page/LoginForm";
+export { default as NavBar } from "./common/NavBar";
+export { default as ProgressBar } from "./common/ProgressBar";
+export { default as ProjectsList } from "./portal/ProjectsList";
+export { default as SamplePreview } from "./portal/SamplePreview";
+export { default as TagDiffPreview } from "./portal/TagDiffPreview";
+export { default as TagPreview } from "./portal/TagPreview";
+export { default as Welcome } from "./activity/Welcome";
