+++ react/src/components/views/index.js
@@ -1,15 +1,17 @@
-export { default as AboutView } from "./AboutView";
-export { default as ActivitiesInfiniteView } from "./ActivitiesInfiniteView";
-export { default as ActivitiesPlayerView } from "./ActivitiesPlayerView";
-export { default as ActivitiesTurkView } from "./ActivitiesTurkView";
-export { default as CareersView } from "./CareersView";
-export { default as HomeView } from "./HomeView";
-export { default as NewProjectView } from "./NewProjectView";
-export { default as NotFoundView } from "./NotFoundView";
-export { default as PrivacyView } from "./PrivacyView";
-export { default as ProjectsListView } from "./ProjectsListView";
-export { default as ProjectView } from "./ProjectView";
-export { default as SamplesView } from "./SamplesView";
-export { default as TagDetailView } from "./TagDetailView";
-export { default as TagDiffView } from "./TagDiffView";
-export { default as TermsView } from "./TermsView";
+export { default as ActivitiesInfiniteView } from "./activity/ActivitiesInfiniteView";
+export { default as ActivitiesPlayerView } from "./activity/ActivitiesPlayerView";
+export { default as ActivitiesTurkView } from "./activity/ActivitiesTurkView";
+
+export { default as AboutView } from "./page/AboutView";
+export { default as CareersView } from "./page/CareersView";
+export { default as HomeView } from "./page/HomeView";
+export { default as NotFoundView } from "./page/NotFoundView";
+export { default as PrivacyView } from "./page/PrivacyView";
+export { default as TermsView } from "./page/TermsView";
+
+export { default as NewProjectView } from "./portal/NewProjectView";
+export { default as ProjectsListView } from "./portal/ProjectsListView";
+export { default as ProjectView } from "./portal/ProjectView";
+export { default as SamplesView } from "./portal/SamplesView";
+export { default as TagDetailView } from "./portal/TagDetailView";
+export { default as TagDiffView } from "./portal/TagDiffView";
