+++ react/src/components/views/page/CareersView.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
-import { Footer, NavBar } from "../elements";
+import { Footer, NavBar } from "../../elements";
 
-import pageStyles from "../../styles/page.module.scss";
+import pageStyles from "../../../styles/page.module.scss";
 
 class CareersView extends Component<any, any> {
     public render() {
