+++ react/src/components/views/page/HomeView.module.scss
@@ -1,5 +1,5 @@
 .tagbullLogo {
-  background-image: url("../../tagbull.svg");
+  background-image: url("../../../tagbull.svg");
   background-repeat: no-repeat;
   background-size: contain;
 }
@@ -59,7 +59,7 @@
 }
 
 .graphic1 {
-  background-image: url("../../images/graphic1.svg");
+  background-image: url("../../../images/graphic1.svg");
 }
 
 @media only screen and (max-width: 650px) {
