+++ react/src/components/views/page/TermsView.tsx
@@ -1,8 +1,8 @@
 /* tslint:disable */
 import React, { Component } from "react";
-import { Footer, NavBar } from "../elements";
+import { Footer, NavBar } from "../../elements";
 
-import pageStyles from "../../styles/page.module.scss";
+import pageStyles from "../../../styles/page.module.scss";
 
 class TermsView extends Component<any, any> {
     public render() {
