+++ react/src/components/views/portal/NewProjectView.module.scss
